change name, delete notice
