+++ EnvironmentChecker/Constraints/PhpConstraint.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace RequiredVersionDisabler\Constraints;
+namespace EnvironmentChecker\Constraints;
 
-class WordPressConstraint extends Constraint
+class PhpConstraint extends Constraint
 {
 
     /**
@@ -18,27 +18,24 @@ class WordPressConstraint extends Constraint
     }
 
     /**
-     * Check if the installation has the required WP version
+     * Check if the installation has the required PHP version
      * show notice if not.
      *
      * @return bool|mixed
      */
     public function check()
     {
-        $WPCurrentVersion = get_bloginfo('version');
-
-        $isWordPressVersionCompatible = version_compare(
-            $WPCurrentVersion,
+        $isPhpVersionCompatible = version_compare(
+            PHP_VERSION,
             $this->requiredVersion,
             '>='
         );
-
-        if (!$isWordPressVersionCompatible) {
+        if (!$isPhpVersionCompatible) {
             $message = sprintf(
                 esc_html__(
-                    '%1$s%3$s%2$s: Plugin disabled. WordPress version has to be '
+                    '%1$s%3$s%2$s: Plugin disabled. PHP version has to be '
                     . $this->requiredVersion
-                    . ' or higher. Please update your WordPress version',
+                    . ' or higher. Please update your PHP version',
                     'mollie-payments-for-woocommerce'
                 ),
                 '<strong>',
@@ -51,5 +48,4 @@ class WordPressConstraint extends Constraint
 
         return true;
     }
-
 }
